refactor viz methods
